requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
scikit-learn
opencv-python-headless
numpy
Pillow
joblib
streamlit
